+++ waspc/cli/src/Wasp/Cli/Command/Call.hs
@@ -13,6 +13,7 @@ data Call
   | Build
   | BuildStart Arguments
   | Version
+  | Doctor
   | Telemetry
   | Deps
   | Dockerfile
@@ -25,5 +26,6 @@ data Call
   | Deploy Arguments -- deploy cmd passthrough args
   | Test Arguments -- "client" | "server", then test cmd passthrough args
   | Unknown Arguments -- all args
+  deriving (Eq)
 
 type Arguments = [String]